requirements.txt
fastapi
uvicorn
chromadb
langchain-community
langchain-groq
langchain-text-splitters
python-dotenv
requests
pypdf
pgvector